organize HashList
