+++ src/StateSmith/compiler/NamedVertexMap.cs
@@ -1,4 +1,5 @@
-﻿using StateSmith.Compiling;
+﻿using StateSmith.Common;
+using StateSmith.Compiling;
 using System.Linq;
 
 #nullable enable
